Remove "square.sort_manifests" in favour of "manio.sort_manifests"
