+++ square/square.py
@@ -1,6 +1,6 @@
 import json
 import logging
-from typing import Collection, Dict, Iterable, List, Optional, Set, Tuple
+from typing import Iterable, Optional, Set, Tuple
 
 import colorama
 import jsonpatch
@@ -12,49 +12,12 @@ import yaml
 from colorlog import ColoredFormatter
 from square.dtypes import (
     DeltaCreate, DeltaDelete, DeltaPatch, DeploymentPlan, DeploymentPlanMeta,
-    Filepath, GroupBy, JsonPatch, K8sConfig, LocalManifestLists, MetaManifest,
-    Selectors, ServerManifests,
+    Filepath, GroupBy, JsonPatch, K8sConfig, MetaManifest, Selectors,
+    ServerManifests,
 )
 
 # Convenience: global logger instance to avoid repetitive code.
 logit = logging.getLogger("square")
-
-
-def sort_manifests(
-        kinds_order: Collection[str],
-        file_manifests: LocalManifestLists,
-) -> Tuple[Dict[Filepath, Collection[MetaManifest]], bool]:
-
-    out: Dict[Filepath, Collection[MetaManifest]] = {}
-    for fname, manifests in file_manifests.items():
-        # Verify that this file contains only supported resource kinds.
-        kinds = {meta.kind for meta, _ in manifests}
-        delta = kinds - set(kinds_order)
-        if len(delta) > 0:
-            logit.error(f"Found unsupported K8s resources in {fname}: {delta}")
-            return ({}, True)
-        del kinds, delta
-
-        # Group the manifests by their "kind", sort each group by their
-        # `MetaManifest` and compile a new list of grouped and sorted
-        # manifests.
-        man_sorted: List[dict] = []
-        for kind in kinds_order:
-            # All manifests of type eg "Service".
-            tmp = [_ for _ in manifests if _[0].kind == kind]
-
-            # Append the manifests ordered by their MetaManifest.
-            man_sorted += sorted(tmp, key=lambda _: _[0])
-        assert len(man_sorted) == len(manifests)
-
-        # Drop the MetaManifest, ie
-        # Dict[Filepath:Tuple[MetaManifest, manifest]] -> Dict[Filepath:manifest]
-        man_clean = [manifest for _, manifest in man_sorted]
-
-        # Assign the grouped and sorted list of manifests to the output dict.
-        out[fname] = man_clean
-
-    return out, False
 
 
 def make_patch(
